consolidated popup, timer, and getSnitchedOn
